+++ CHANGES.md
@@ -4,6 +4,10 @@ moodle-tool_selfsignuphardlifecycle
 Changes
 -------
 
+### Unreleased
+
+* 2022-04-28 - Feature: Add possibility to override individual users.
+
 ### v3.9-r1
 
 * 2022-03-01 - Initial version.